+++ test/runtests.jl
@@ -15,6 +15,7 @@ include("erlang_loss.jl")
 # All available test files
 all_tests = [
     "test_package.jl",
+    "test_autodiff.jl",
     "test_combinednr.jl",
     "test_context.jl",
     "test_delayed_state.jl",